requirements.txt
pymongo==4.3.3
python-dotenv
requests==2.28.2
selenium==4.8.2
webdriver-manager==3.8.5
pillow
#apache-airflow[amazon]
